+++ src/librarymanagementsystem/dao/LibraryRepository.java
@@ -6,10 +6,10 @@ import java.util.List;
 
 public interface LibraryRepository {
     List<Library> findAll();
-    Library findById(Long id);
+    Library findById(int id);
     List<Library> findByName(String name);
     void create(Library library);
     Library update(Library library);
     void remove(Library library);
-    String viewInfo(Long id);
+    String viewInfo(int id);
 }
